requirements.txt
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
joblib==1.4.0
kiwisolver==1.4.5
llvmlite==0.42.0
matplotlib==3.8.4
numba==0.59.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pynndescent==0.5.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
threadpoolctl==3.4.0
tqdm==4.66.2
umap-learn==0.5.6
